requirements.txt
alembic==1.7.4
backports.zoneinfo==0.2.1
blinker==1.4
click==8.0.3
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
psycopg==3.0.1
psycopg2==2.9.1
python-dotenv==0.19.1
SQLAlchemy==1.4.26
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
